make metrics filter lookup fetches non-blocking
