+++ ui/apps/dashboard/src/gql/gql.ts
@@ -94,9 +94,9 @@ const documents = {
     "\n  query GetFunctionVersionNumber($slug: String!, $environmentID: ID!) {\n    workspace(id: $environmentID) {\n      workflow: workflowBySlug(slug: $slug) {\n        id\n        isPaused\n        name\n        archivedAt\n        current {\n          version\n        }\n        previous {\n          version\n        }\n      }\n    }\n  }\n": types.GetFunctionVersionNumberDocument,
     "\n  mutation PauseFunction($fnID: ID!, $cancelRunning: Boolean) {\n    pauseFunction(fnID: $fnID, cancelRunning: $cancelRunning) {\n      id\n    }\n  }\n": types.PauseFunctionDocument,
     "\n  mutation UnpauseFunction($fnID: ID!) {\n    unpauseFunction(fnID: $fnID) {\n      id\n    }\n  }\n": types.UnpauseFunctionDocument,
+    "\n  query MetricsLookups($envSlug: String!, $page: Int, $pageSize: Int) {\n    envBySlug(slug: $envSlug) {\n      apps {\n        externalID\n        id\n        name\n        isArchived\n      }\n      workflows @paginated(perPage: $pageSize, page: $page) {\n        data {\n          name\n          id\n          slug\n        }\n        page {\n          page\n          totalPages\n          perPage\n        }\n      }\n    }\n  }\n": types.MetricsLookupsDocument,
     "\n  query FunctionStatusMetrics(\n    $workspaceId: ID!\n    $from: Time!\n    $functionIDs: [UUID!]\n    $appIDs: [UUID!]\n    $until: Time\n    $scope: MetricsScope!\n  ) {\n    workspace(id: $workspaceId) {\n      scheduled: scopedMetrics(\n        filter: {\n          name: \"function_run_scheduled_total\"\n          scope: $scope\n          from: $from\n          functionIDs: $functionIDs\n          appIDs: $appIDs\n          until: $until\n        }\n      ) {\n        metrics {\n          id\n          data {\n            value\n            bucket\n          }\n        }\n      }\n    }\n    workspace(id: $workspaceId) {\n      started: scopedMetrics(\n        filter: {\n          name: \"function_run_started_total\"\n          scope: $scope\n          from: $from\n          functionIDs: $functionIDs\n          appIDs: $appIDs\n          until: $until\n        }\n      ) {\n        metrics {\n          id\n          data {\n            value\n            bucket\n          }\n        }\n      }\n    }\n    workspace(id: $workspaceId) {\n      completed: scopedMetrics(\n        filter: {\n          name: \"function_run_ended_total\"\n          scope: $scope\n          groupBy: \"status\"\n          from: $from\n          functionIDs: $functionIDs\n          appIDs: $appIDs\n          until: $until\n        }\n      ) {\n        metrics {\n          id\n          tagName\n          tagValue\n          data {\n            value\n            bucket\n          }\n        }\n      }\n    }\n    workspace(id: $workspaceId) {\n      completedByFunction: scopedMetrics(\n        filter: {\n          name: \"function_run_ended_total\"\n          scope: FN\n          groupBy: \"status\"\n          from: $from\n          functionIDs: $functionIDs\n          appIDs: $appIDs\n          until: $until\n        }\n      ) {\n        metrics {\n          id\n          tagName\n          tagValue\n          data {\n            value\n            bucket\n          }\n        }\n      }\n    }\n    workspace(id: $workspaceId) {\n      totals: scopedFunctionStatus(\n        filter: {\n          name: \"function_run_scheduled_total\"\n          scope: FN\n          from: $from\n          functionIDs: $functionIDs\n          appIDs: $appIDs\n          until: $until\n        }\n      ) {\n        queued\n        running\n        completed\n        failed\n        cancelled\n        cancelled\n        skipped\n      }\n    }\n  }\n": types.FunctionStatusMetricsDocument,
     "\n  query VolumeMetrics(\n    $workspaceId: ID!\n    $from: Time!\n    $functionIDs: [UUID!]\n    $appIDs: [UUID!]\n    $until: Time\n    $scope: MetricsScope!\n  ) {\n    workspace(id: $workspaceId) {\n      runsThroughput: scopedMetrics(\n        filter: {\n          name: \"function_run_ended_total\"\n          scope: $scope\n          from: $from\n          functionIDs: $functionIDs\n          appIDs: $appIDs\n          until: $until\n        }\n      ) {\n        metrics {\n          id\n          tagName\n          tagValue\n          data {\n            value\n            bucket\n          }\n        }\n      }\n    }\n    workspace(id: $workspaceId) {\n      sdkThroughputEnded: scopedMetrics(\n        filter: {\n          name: \"sdk_req_ended_total\"\n          scope: $scope\n          from: $from\n          functionIDs: $functionIDs\n          appIDs: $appIDs\n          until: $until\n        }\n      ) {\n        metrics {\n          id\n          tagName\n          tagValue\n          data {\n            value\n            bucket\n          }\n        }\n      }\n    }\n    workspace(id: $workspaceId) {\n      sdkThroughputStarted: scopedMetrics(\n        filter: {\n          name: \"sdk_req_started_total\"\n          scope: $scope\n          from: $from\n          functionIDs: $functionIDs\n          appIDs: $appIDs\n          until: $until\n        }\n      ) {\n        metrics {\n          id\n          tagName\n          tagValue\n          data {\n            value\n            bucket\n          }\n        }\n      }\n    }\n    workspace(id: $workspaceId) {\n      sdkThroughputScheduled: scopedMetrics(\n        filter: {\n          name: \"sdk_req_scheduled_total\"\n          scope: $scope\n          from: $from\n          functionIDs: $functionIDs\n          appIDs: $appIDs\n          until: $until\n        }\n      ) {\n        metrics {\n          id\n          tagName\n          tagValue\n          data {\n            value\n            bucket\n          }\n        }\n      }\n    }\n    workspace(id: $workspaceId) {\n      stepThroughput: scopedMetrics(\n        filter: {\n          name: \"step_output_bytes_total\"\n          scope: $scope\n          from: $from\n          functionIDs: $functionIDs\n          appIDs: $appIDs\n          until: $until\n        }\n      ) {\n        metrics {\n          id\n          tagName\n          tagValue\n          data {\n            value\n            bucket\n          }\n        }\n      }\n    }\n    workspace(id: $workspaceId) {\n      backlog: scopedMetrics(\n        filter: {\n          name: \"steps_scheduled\"\n          scope: $scope\n          from: $from\n          functionIDs: $functionIDs\n          appIDs: $appIDs\n          until: $until\n        }\n      ) {\n        metrics {\n          id\n          tagName\n          tagValue\n          data {\n            value\n            bucket\n          }\n        }\n      }\n    }\n    workspace(id: $workspaceId) {\n      stepRunning: scopedMetrics(\n        filter: {\n          name: \"steps_running\"\n          scope: $scope\n          from: $from\n          functionIDs: $functionIDs\n          appIDs: $appIDs\n          until: $until\n        }\n      ) {\n        metrics {\n          id\n          tagName\n          tagValue\n          data {\n            value\n            bucket\n          }\n        }\n      }\n    }\n    workspace(id: $workspaceId) {\n      concurrency: scopedMetrics(\n        filter: {\n          name: \"concurrency_limit_reached_total\"\n          scope: $scope\n          from: $from\n          functionIDs: $functionIDs\n          appIDs: $appIDs\n          until: $until\n        }\n      ) {\n        metrics {\n          id\n          tagName\n          tagValue\n          data {\n            value\n            bucket\n          }\n        }\n      }\n    }\n  }\n": types.VolumeMetricsDocument,
-    "\n  query MetricsLookups($envSlug: String!, $page: Int, $pageSize: Int) {\n    envBySlug(slug: $envSlug) {\n      apps {\n        externalID\n        id\n        name\n        isArchived\n      }\n      workflows @paginated(perPage: $pageSize, page: $page) {\n        data {\n          name\n          id\n          slug\n        }\n        page {\n          page\n          totalPages\n          perPage\n        }\n      }\n    }\n  }\n": types.MetricsLookupsDocument,
     "\n  query GetGlobalSearch($opts: SearchInput!) {\n    account {\n      search(opts: $opts) {\n        results {\n          env {\n            name\n            id\n            type\n            slug\n          }\n          kind\n          value {\n            ... on ArchivedEvent {\n              id\n              name\n            }\n            ... on FunctionRun {\n              id\n              functionID: workflowID\n              startedAt\n            }\n          }\n        }\n      }\n    }\n  }\n": types.GetGlobalSearchDocument,
     "\n  query GetFunctionSlug($environmentID: ID!, $functionID: ID!) {\n    environment: workspace(id: $environmentID) {\n      function: workflow(id: $functionID) {\n        slug\n        name\n      }\n    }\n  }\n": types.GetFunctionSlugDocument,
     "\n  mutation SyncOnboardingApp($appURL: String!, $envID: UUID!) {\n    syncNewApp(appURL: $appURL, envID: $envID) {\n      app {\n        externalID\n        id\n      }\n      error {\n        code\n        data\n        message\n      }\n    }\n  }\n": types.SyncOnboardingAppDocument,
@@ -468,15 +468,15 @@ export function graphql(source: "\n  mutation UnpauseFunction($fnID: ID!) {\n   
 /**
  * The graphql function is used to parse GraphQL queries into a document that can be used by GraphQL clients.
  */
+export function graphql(source: "\n  query MetricsLookups($envSlug: String!, $page: Int, $pageSize: Int) {\n    envBySlug(slug: $envSlug) {\n      apps {\n        externalID\n        id\n        name\n        isArchived\n      }\n      workflows @paginated(perPage: $pageSize, page: $page) {\n        data {\n          name\n          id\n          slug\n        }\n        page {\n          page\n          totalPages\n          perPage\n        }\n      }\n    }\n  }\n"): (typeof documents)["\n  query MetricsLookups($envSlug: String!, $page: Int, $pageSize: Int) {\n    envBySlug(slug: $envSlug) {\n      apps {\n        externalID\n        id\n        name\n        isArchived\n      }\n      workflows @paginated(perPage: $pageSize, page: $page) {\n        data {\n          name\n          id\n          slug\n        }\n        page {\n          page\n          totalPages\n          perPage\n        }\n      }\n    }\n  }\n"];
+/**
+ * The graphql function is used to parse GraphQL queries into a document that can be used by GraphQL clients.
+ */
 export function graphql(source: "\n  query FunctionStatusMetrics(\n    $workspaceId: ID!\n    $from: Time!\n    $functionIDs: [UUID!]\n    $appIDs: [UUID!]\n    $until: Time\n    $scope: MetricsScope!\n  ) {\n    workspace(id: $workspaceId) {\n      scheduled: scopedMetrics(\n        filter: {\n          name: \"function_run_scheduled_total\"\n          scope: $scope\n          from: $from\n          functionIDs: $functionIDs\n          appIDs: $appIDs\n          until: $until\n        }\n      ) {\n        metrics {\n          id\n          data {\n            value\n            bucket\n          }\n        }\n      }\n    }\n    workspace(id: $workspaceId) {\n      started: scopedMetrics(\n        filter: {\n          name: \"function_run_started_total\"\n          scope: $scope\n          from: $from\n          functionIDs: $functionIDs\n          appIDs: $appIDs\n          until: $until\n        }\n      ) {\n        metrics {\n          id\n          data {\n            value\n            bucket\n          }\n        }\n      }\n    }\n    workspace(id: $workspaceId) {\n      completed: scopedMetrics(\n        filter: {\n          name: \"function_run_ended_total\"\n          scope: $scope\n          groupBy: \"status\"\n          from: $from\n          functionIDs: $functionIDs\n          appIDs: $appIDs\n          until: $until\n        }\n      ) {\n        metrics {\n          id\n          tagName\n          tagValue\n          data {\n            value\n            bucket\n          }\n        }\n      }\n    }\n    workspace(id: $workspaceId) {\n      completedByFunction: scopedMetrics(\n        filter: {\n          name: \"function_run_ended_total\"\n          scope: FN\n          groupBy: \"status\"\n          from: $from\n          functionIDs: $functionIDs\n          appIDs: $appIDs\n          until: $until\n        }\n      ) {\n        metrics {\n          id\n          tagName\n          tagValue\n          data {\n            value\n            bucket\n          }\n        }\n      }\n    }\n    workspace(id: $workspaceId) {\n      totals: scopedFunctionStatus(\n        filter: {\n          name: \"function_run_scheduled_total\"\n          scope: FN\n          from: $from\n          functionIDs: $functionIDs\n          appIDs: $appIDs\n          until: $until\n        }\n      ) {\n        queued\n        running\n        completed\n        failed\n        cancelled\n        cancelled\n        skipped\n      }\n    }\n  }\n"): (typeof documents)["\n  query FunctionStatusMetrics(\n    $workspaceId: ID!\n    $from: Time!\n    $functionIDs: [UUID!]\n    $appIDs: [UUID!]\n    $until: Time\n    $scope: MetricsScope!\n  ) {\n    workspace(id: $workspaceId) {\n      scheduled: scopedMetrics(\n        filter: {\n          name: \"function_run_scheduled_total\"\n          scope: $scope\n          from: $from\n          functionIDs: $functionIDs\n          appIDs: $appIDs\n          until: $until\n        }\n      ) {\n        metrics {\n          id\n          data {\n            value\n            bucket\n          }\n        }\n      }\n    }\n    workspace(id: $workspaceId) {\n      started: scopedMetrics(\n        filter: {\n          name: \"function_run_started_total\"\n          scope: $scope\n          from: $from\n          functionIDs: $functionIDs\n          appIDs: $appIDs\n          until: $until\n        }\n      ) {\n        metrics {\n          id\n          data {\n            value\n            bucket\n          }\n        }\n      }\n    }\n    workspace(id: $workspaceId) {\n      completed: scopedMetrics(\n        filter: {\n          name: \"function_run_ended_total\"\n          scope: $scope\n          groupBy: \"status\"\n          from: $from\n          functionIDs: $functionIDs\n          appIDs: $appIDs\n          until: $until\n        }\n      ) {\n        metrics {\n          id\n          tagName\n          tagValue\n          data {\n            value\n            bucket\n          }\n        }\n      }\n    }\n    workspace(id: $workspaceId) {\n      completedByFunction: scopedMetrics(\n        filter: {\n          name: \"function_run_ended_total\"\n          scope: FN\n          groupBy: \"status\"\n          from: $from\n          functionIDs: $functionIDs\n          appIDs: $appIDs\n          until: $until\n        }\n      ) {\n        metrics {\n          id\n          tagName\n          tagValue\n          data {\n            value\n            bucket\n          }\n        }\n      }\n    }\n    workspace(id: $workspaceId) {\n      totals: scopedFunctionStatus(\n        filter: {\n          name: \"function_run_scheduled_total\"\n          scope: FN\n          from: $from\n          functionIDs: $functionIDs\n          appIDs: $appIDs\n          until: $until\n        }\n      ) {\n        queued\n        running\n        completed\n        failed\n        cancelled\n        cancelled\n        skipped\n      }\n    }\n  }\n"];
 /**
  * The graphql function is used to parse GraphQL queries into a document that can be used by GraphQL clients.
  */
 export function graphql(source: "\n  query VolumeMetrics(\n    $workspaceId: ID!\n    $from: Time!\n    $functionIDs: [UUID!]\n    $appIDs: [UUID!]\n    $until: Time\n    $scope: MetricsScope!\n  ) {\n    workspace(id: $workspaceId) {\n      runsThroughput: scopedMetrics(\n        filter: {\n          name: \"function_run_ended_total\"\n          scope: $scope\n          from: $from\n          functionIDs: $functionIDs\n          appIDs: $appIDs\n          until: $until\n        }\n      ) {\n        metrics {\n          id\n          tagName\n          tagValue\n          data {\n            value\n            bucket\n          }\n        }\n      }\n    }\n    workspace(id: $workspaceId) {\n      sdkThroughputEnded: scopedMetrics(\n        filter: {\n          name: \"sdk_req_ended_total\"\n          scope: $scope\n          from: $from\n          functionIDs: $functionIDs\n          appIDs: $appIDs\n          until: $until\n        }\n      ) {\n        metrics {\n          id\n          tagName\n          tagValue\n          data {\n            value\n            bucket\n          }\n        }\n      }\n    }\n    workspace(id: $workspaceId) {\n      sdkThroughputStarted: scopedMetrics(\n        filter: {\n          name: \"sdk_req_started_total\"\n          scope: $scope\n          from: $from\n          functionIDs: $functionIDs\n          appIDs: $appIDs\n          until: $until\n        }\n      ) {\n        metrics {\n          id\n          tagName\n          tagValue\n          data {\n            value\n            bucket\n          }\n        }\n      }\n    }\n    workspace(id: $workspaceId) {\n      sdkThroughputScheduled: scopedMetrics(\n        filter: {\n          name: \"sdk_req_scheduled_total\"\n          scope: $scope\n          from: $from\n          functionIDs: $functionIDs\n          appIDs: $appIDs\n          until: $until\n        }\n      ) {\n        metrics {\n          id\n          tagName\n          tagValue\n          data {\n            value\n            bucket\n          }\n        }\n      }\n    }\n    workspace(id: $workspaceId) {\n      stepThroughput: scopedMetrics(\n        filter: {\n          name: \"step_output_bytes_total\"\n          scope: $scope\n          from: $from\n          functionIDs: $functionIDs\n          appIDs: $appIDs\n          until: $until\n        }\n      ) {\n        metrics {\n          id\n          tagName\n          tagValue\n          data {\n            value\n            bucket\n          }\n        }\n      }\n    }\n    workspace(id: $workspaceId) {\n      backlog: scopedMetrics(\n        filter: {\n          name: \"steps_scheduled\"\n          scope: $scope\n          from: $from\n          functionIDs: $functionIDs\n          appIDs: $appIDs\n          until: $until\n        }\n      ) {\n        metrics {\n          id\n          tagName\n          tagValue\n          data {\n            value\n            bucket\n          }\n        }\n      }\n    }\n    workspace(id: $workspaceId) {\n      stepRunning: scopedMetrics(\n        filter: {\n          name: \"steps_running\"\n          scope: $scope\n          from: $from\n          functionIDs: $functionIDs\n          appIDs: $appIDs\n          until: $until\n        }\n      ) {\n        metrics {\n          id\n          tagName\n          tagValue\n          data {\n            value\n            bucket\n          }\n        }\n      }\n    }\n    workspace(id: $workspaceId) {\n      concurrency: scopedMetrics(\n        filter: {\n          name: \"concurrency_limit_reached_total\"\n          scope: $scope\n          from: $from\n          functionIDs: $functionIDs\n          appIDs: $appIDs\n          until: $until\n        }\n      ) {\n        metrics {\n          id\n          tagName\n          tagValue\n          data {\n            value\n            bucket\n          }\n        }\n      }\n    }\n  }\n"): (typeof documents)["\n  query VolumeMetrics(\n    $workspaceId: ID!\n    $from: Time!\n    $functionIDs: [UUID!]\n    $appIDs: [UUID!]\n    $until: Time\n    $scope: MetricsScope!\n  ) {\n    workspace(id: $workspaceId) {\n      runsThroughput: scopedMetrics(\n        filter: {\n          name: \"function_run_ended_total\"\n          scope: $scope\n          from: $from\n          functionIDs: $functionIDs\n          appIDs: $appIDs\n          until: $until\n        }\n      ) {\n        metrics {\n          id\n          tagName\n          tagValue\n          data {\n            value\n            bucket\n          }\n        }\n      }\n    }\n    workspace(id: $workspaceId) {\n      sdkThroughputEnded: scopedMetrics(\n        filter: {\n          name: \"sdk_req_ended_total\"\n          scope: $scope\n          from: $from\n          functionIDs: $functionIDs\n          appIDs: $appIDs\n          until: $until\n        }\n      ) {\n        metrics {\n          id\n          tagName\n          tagValue\n          data {\n            value\n            bucket\n          }\n        }\n      }\n    }\n    workspace(id: $workspaceId) {\n      sdkThroughputStarted: scopedMetrics(\n        filter: {\n          name: \"sdk_req_started_total\"\n          scope: $scope\n          from: $from\n          functionIDs: $functionIDs\n          appIDs: $appIDs\n          until: $until\n        }\n      ) {\n        metrics {\n          id\n          tagName\n          tagValue\n          data {\n            value\n            bucket\n          }\n        }\n      }\n    }\n    workspace(id: $workspaceId) {\n      sdkThroughputScheduled: scopedMetrics(\n        filter: {\n          name: \"sdk_req_scheduled_total\"\n          scope: $scope\n          from: $from\n          functionIDs: $functionIDs\n          appIDs: $appIDs\n          until: $until\n        }\n      ) {\n        metrics {\n          id\n          tagName\n          tagValue\n          data {\n            value\n            bucket\n          }\n        }\n      }\n    }\n    workspace(id: $workspaceId) {\n      stepThroughput: scopedMetrics(\n        filter: {\n          name: \"step_output_bytes_total\"\n          scope: $scope\n          from: $from\n          functionIDs: $functionIDs\n          appIDs: $appIDs\n          until: $until\n        }\n      ) {\n        metrics {\n          id\n          tagName\n          tagValue\n          data {\n            value\n            bucket\n          }\n        }\n      }\n    }\n    workspace(id: $workspaceId) {\n      backlog: scopedMetrics(\n        filter: {\n          name: \"steps_scheduled\"\n          scope: $scope\n          from: $from\n          functionIDs: $functionIDs\n          appIDs: $appIDs\n          until: $until\n        }\n      ) {\n        metrics {\n          id\n          tagName\n          tagValue\n          data {\n            value\n            bucket\n          }\n        }\n      }\n    }\n    workspace(id: $workspaceId) {\n      stepRunning: scopedMetrics(\n        filter: {\n          name: \"steps_running\"\n          scope: $scope\n          from: $from\n          functionIDs: $functionIDs\n          appIDs: $appIDs\n          until: $until\n        }\n      ) {\n        metrics {\n          id\n          tagName\n          tagValue\n          data {\n            value\n            bucket\n          }\n        }\n      }\n    }\n    workspace(id: $workspaceId) {\n      concurrency: scopedMetrics(\n        filter: {\n          name: \"concurrency_limit_reached_total\"\n          scope: $scope\n          from: $from\n          functionIDs: $functionIDs\n          appIDs: $appIDs\n          until: $until\n        }\n      ) {\n        metrics {\n          id\n          tagName\n          tagValue\n          data {\n            value\n            bucket\n          }\n        }\n      }\n    }\n  }\n"];
-/**
- * The graphql function is used to parse GraphQL queries into a document that can be used by GraphQL clients.
- */
-export function graphql(source: "\n  query MetricsLookups($envSlug: String!, $page: Int, $pageSize: Int) {\n    envBySlug(slug: $envSlug) {\n      apps {\n        externalID\n        id\n        name\n        isArchived\n      }\n      workflows @paginated(perPage: $pageSize, page: $page) {\n        data {\n          name\n          id\n          slug\n        }\n        page {\n          page\n          totalPages\n          perPage\n        }\n      }\n    }\n  }\n"): (typeof documents)["\n  query MetricsLookups($envSlug: String!, $page: Int, $pageSize: Int) {\n    envBySlug(slug: $envSlug) {\n      apps {\n        externalID\n        id\n        name\n        isArchived\n      }\n      workflows @paginated(perPage: $pageSize, page: $page) {\n        data {\n          name\n          id\n          slug\n        }\n        page {\n          page\n          totalPages\n          perPage\n        }\n      }\n    }\n  }\n"];
 /**
  * The graphql function is used to parse GraphQL queries into a document that can be used by GraphQL clients.
  */
